these changes restructure the game into classes and begin to add event listeners. There is still a lot of work to be done but I wanted to commit an update of how it was evolving.
I don't think I will need a node server or even an apache server.
I'm thinking of hosting it just as is on github.

I am using the type=module attribute for html script tag. For
compatability I will probably add a 'defer nomodule' and have a
webpack version that is able to load in it's place.

More updates to come this week.