Restructured the Gameplay.js file extensively and finally completed the clear square functions. I'm sure there is a better way to do this but I finally found a method that worked for my brain. It take it in pieces. First, I find and clear the blank squares. Then, I store that array and pass it to a clear numbers function. This then clears the perimeter of the blank squares, which are the numbered squares. This brings the board to a point where all touching blank squares and their perimeter of numbered squares are cleared upon click of a blank square.
I also cleaned up the code so no matter if we are talking about a
parameter, variable, object, whatever etc. the row is always listed
first. I found it hard to diagnose my own code when I was flip flopping
row and col when listing them in their many forms. Keeping them always
row first and col 2nd help diagnose the code and move it forward.

Things that are still missing...
* Gameover needs to actually be Gameover ...
* Bomb counter.. which I'd like to actually be number of squares left to
uncover, less the amount of bombs.
* a cool timer to keep time
* consider support contextmenu (right click) flags and also add that
logic as a stretch goal

I kind of like the idea of just clearing all the squares without use of
flags but I'm not a minesweeper guru so I'm sure they have their place.

That's all the updates for now. More commits soon.