over the weekend I made some major updates to mobile responsiveness / css restructuring. I also and added the leaderboard structure and and gametimer and squares remaining counter to the HUD.
the game is currenting using localStorage to store playerName,
datePlayed and timeFinished variables.

Last step I see is updating the leaderboard upon victory.

The plan is to only hold the last 5-10 victories and list them in order
of best time. These will also be stored in localStorage obviously so
they are retained through a refresh.

More updates soon.