leaderBoard and ranking functionality has been added. It all works fairly well at this point. Again, I'm sure there was an easier way to do this but I feel it's working well now and I can revisit better idea to handle this later if needed.
overall, still a few items that need to done and or looked at.

- Add game board size to leaderboard info e.g. Small

- style text so it shows up cleaner on leaderBoard

- fix small alignment issues with #board on gamestart

- when typing in a name on mobile, zoom back out after focus
function so it’s not zoomed in when game starts (on mobile)

- i also noticed very seldom, like maybe 1 in 25-30 games, a gameover
gamestats seems to make it to the leaderBoard. Not sure what's
happening there. might be how I am testing in VScode with browser sync
extension. one thing I do want to think about is how the end of the game
is calculated. Right now it uses square visibility which may be the
reason for the occational incorrect leaderBoard writing but I'm not
sure. Going to keep testing.